requirements.txt
hf-xet
tqdm
transformers
pyyaml
llama-cpp-python
pynvml
